requirements.txt
flask==2.0.1
flask-cors==3.0.10
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
python-whois==0.7.3
requests==2.26.0
beautifulsoup4==4.9.3
tldextract==3.1.2 